refactor: make file and module names consistent in preparation for introducing zeitwerk
