+++ lib/pact_broker/app.rb
@@ -23,7 +23,6 @@ require "rack/pact_broker/application_context"
 require "sucker_punch"
 require "pact_broker/api/middleware/configuration"
 require "pact_broker/api/middleware/basic_auth"
-require "pact_broker/config/basic_auth_configuration"
 require "pact_broker/api/authorization/resource_access_policy"
 require "pact_broker/api/middleware/http_debug_logs"
 require "pact_broker/application_context"
