+++ lib/pact_broker/config/runtime_configuration.rb
@@ -3,7 +3,7 @@ require "pact_broker/config/runtime_configuration_logging_methods"
 require "pact_broker/config/runtime_configuration_database_methods"
 require "pact_broker/config/runtime_configuration_coercion_methods"
 require "pact_broker/version"
-require "pact_broker/config/basic_auth_configuration"
+require "pact_broker/config/runtime_configuration_basic_auth_methods"
 require "pact_broker/string_refinements"
 require "pact_broker/hash_refinements"
 require "pact_broker/error"
