+++ lib/pact_broker/pacts/pact_publication.rb
@@ -3,7 +3,7 @@ require "pact_broker/dataset"
 require "pact_broker/domain/pact"
 require "pact_broker/pacts/pact_version"
 require "pact_broker/integrations/integration"
-require "pact_broker/tags/head_pact_tags"
+require "pact_broker/tags/head_pact_tag"
 require "pact_broker/pacts/pact_publication_dataset_module"
 require "pact_broker/pacts/pact_publication_wip_dataset_module"
 require "pact_broker/pacts/eager_loaders"
