+++ lib/pact_broker/ui/controllers/clusters.rb
@@ -18,7 +18,7 @@ module PactBroker
         end
 
         get "/" do
-          view_model = ViewDomain::IndexItems.new(pacticipant_service.find_index_items, base_url: base_url)
+          view_model = ViewModel::IndexItems.new(pacticipant_service.find_index_items, base_url: base_url)
           haml "clusters/show", locals: { relationships: view_model, base_url: base_url }, escape_html: true
         end
