+++ lib/pact_broker/ui/controllers/index.rb
@@ -1,4 +1,4 @@
-require "pact_broker/ui/controllers/base_controller"
+require "pact_broker/ui/controllers/base"
 require "pact_broker/ui/view_models/index_items"
 require "pact_broker/string_refinements"
 require "haml"
@@ -33,7 +33,7 @@ module PactBroker
             error_messages << "No pacticipants found matching: \"#{search}\""
           end
 
-          view_index_items = ViewDomain::IndexItems.new(index_items, base_url: base_url)
+          view_index_items = PactBroker::UI::ViewModels::IndexItems.new(index_items, base_url: base_url)
 
           page = tags ? :'index/show-with-tags' : :'index/show'
           locals = {
