+++ lib/pact_broker/ui/controllers/matrix.rb
@@ -1,4 +1,4 @@
-require "pact_broker/ui/controllers/base_controller"
+require "pact_broker/ui/controllers/base"
 require "pact_broker/ui/view_models/matrix_lines"
 require "pact_broker/matrix/unresolved_selector"
 require "pact_broker/matrix/parse_query"
@@ -32,7 +32,7 @@ module PactBroker
               errors = matrix_service.validate_selectors(selectors, options)
               if errors.empty?
                 lines = matrix_service.find(selectors, options)
-                locals[:lines] = PactBroker::UI::ViewDomain::MatrixLines.new(lines, base_url: base_url)
+                locals[:lines] = PactBroker::UI::ViewModels::MatrixLines.new(lines, base_url: base_url)
                 locals[:badge_url] = matrix_badge_url(selectors, lines, base_url)
               else
                 locals[:errors] = errors
@@ -52,7 +52,7 @@ module PactBroker
                       ]
           options = { latestby: "cvpv", limit: 100 }
           lines = matrix_service.find(selectors, options)
-          lines = PactBroker::UI::ViewDomain::MatrixLines.new(lines, base_url: base_url)
+          lines = PactBroker::UI::ViewModels::MatrixLines.new(lines, base_url: base_url)
           locals =  {
                       lines: lines,
                       consumer_name: params[:consumer_name],
