+++ lib/pact_broker/ui/view_models/index_item_branch_head.rb
@@ -4,7 +4,7 @@ require "pact_broker/date_helper"
 
 module PactBroker
   module UI
-    module ViewDomain
+    module ViewModels
       class IndexItemBranchHead
 
         include PactBroker::Api::PactBrokerUrls
