+++ lib/pact_broker/ui/view_models/matrix_lines.rb
@@ -2,12 +2,12 @@ require "pact_broker/ui/view_models/matrix_line"
 
 module PactBroker
   module UI
-    module ViewDomain
+    module ViewModels
       class MatrixLines < Array
 
         def initialize rows, options = {}
           lines = rows.collect do | row |
-            PactBroker::UI::ViewDomain::MatrixLine.new(row, options)
+            PactBroker::UI::ViewModels::MatrixLine.new(row, options)
           end
           super(lines.sort)
         end
