+++ lib/pact_broker/ui/view_models/matrix_released_version.rb
@@ -4,7 +4,7 @@ require "pact_broker/date_helper"
 
 module PactBroker
   module UI
-    module ViewDomain
+    module ViewModels
       class MatrixReleasedVersion
         include PactBroker::Api::PactBrokerUrls
