+++ spec/lib/pact_broker/api/contracts/pacts_for_verification_json_query_schema_combinations_spec.rb
@@ -3,7 +3,7 @@ require "pact_broker/api/contracts/pacts_for_verification_json_query_schema"
 module PactBroker
   module Api
     module Contracts
-      describe PactsForVerificationJSONQuerySchema do
+      describe PactsForVerificationJsonQuerySchema do
         include PactBroker::Test::ApiContractSupport
 
         ALL_PROPERTIES = {
@@ -51,7 +51,7 @@ module PactBroker
 
             it "is valid" do
               params = { consumerVersionSelectors: [selector], providerVersionBranch: "main" }
-              expect(PactsForVerificationJSONQuerySchema.(params).errors).to be_empty
+              expect(PactsForVerificationJsonQuerySchema.(params).errors).to be_empty
             end
 
             extra_keys = ALL_PROPERTIES.keys - valid_key_combination - [:consumer]
@@ -63,11 +63,11 @@ module PactBroker
               describe "with #{selector_with_extra_key}" do
                 if expect_to_be_valid
                   it "is valid" do
-                    expect(PactsForVerificationJSONQuerySchema.(params).errors).to be_empty
+                    expect(PactsForVerificationJsonQuerySchema.(params).errors).to be_empty
                   end
                 else
                   it "is not valid" do
-                    expect(PactsForVerificationJSONQuerySchema.(params).errors).to_not be_empty
+                    expect(PactsForVerificationJsonQuerySchema.(params).errors).to_not be_empty
                   end
                 end
               end
@@ -78,7 +78,7 @@ module PactBroker
                 it "is valid" do
                   params = { consumerVersionSelectors: [selector_with_consumer], providerVersionBranch: "main" }
 
-                  expect(PactsForVerificationJSONQuerySchema.(params).errors.empty?).to be true
+                  expect(PactsForVerificationJsonQuerySchema.(params).errors.empty?).to be true
                 end
               end
             end
